requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlite3
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# AI/ML
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10

# Data processing
pandas==2.1.4
numpy==1.24.3
faker==20.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0

# Web Dashboard
streamlit==1.28.1
plotly==5.17.0
streamlit-chat==0.1.1
